+++ src/main/java/chess/view/PositionParser.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.view;
 
 import chess.domain.board.File;
 import chess.domain.board.Position;
